Move directory helpers to fs mod
