+++ src/lib.rs
@@ -19,6 +19,7 @@ pub mod command;
 pub mod config;
 pub mod ethereum;
 pub mod float_maths;
+pub mod fs;
 pub mod history;
 pub mod jsonrpc;
 pub mod maker;
@@ -31,52 +32,15 @@ pub mod seed;
 pub mod swap;
 pub mod swap_id;
 
-use anyhow::Context;
+#[cfg(all(test, feature = "test-docker"))]
+pub mod test_harness;
+
 use conquer_once::Lazy;
 pub use maker::Maker;
 pub use mid_market_rate::MidMarketRate;
 pub use rate::{Rate, Spread};
 pub use seed::Seed;
-use std::path::{Path, PathBuf};
 pub use swap_id::SwapId;
 
 pub static SECP: Lazy<::bitcoin::secp256k1::Secp256k1<::bitcoin::secp256k1::All>> =
     Lazy::new(::bitcoin::secp256k1::Secp256k1::new);
-
-/// This is to store the configuration and seed files
-// Linux: /home/<user>/.config/nectar/
-// OSX: /Users/<user>/Library/Preferences/nectar/
-fn config_dir() -> Option<PathBuf> {
-    directories::ProjectDirs::from("", "", "nectar")
-        .map(|proj_dirs| proj_dirs.config_dir().to_path_buf())
-}
-
-pub fn default_config_path() -> anyhow::Result<PathBuf> {
-    crate::config_dir()
-        .map(|dir| Path::join(&dir, "config.toml"))
-        .context("Could not generate default configuration path")
-}
-
-/// This is to store the DB
-// Linux: /home/<user>/.local/share/nectar/
-// OSX: /Users/<user>/Library/Application Support/nectar/
-pub fn data_dir() -> Option<std::path::PathBuf> {
-    directories::ProjectDirs::from("", "", "nectar")
-        .map(|proj_dirs| proj_dirs.data_dir().to_path_buf())
-}
-
-pub fn ensure_directory_exists(file: &Path) -> Result<(), std::io::Error> {
-    if let Some(path) = file.parent() {
-        if !path.exists() {
-            tracing::info!(
-                "Parent directory does not exist, creating recursively: {}",
-                file.display()
-            );
-            return std::fs::create_dir_all(path);
-        }
-    }
-    Ok(())
-}
-
-#[cfg(all(test, feature = "test-docker"))]
-pub mod test_harness;
